Add telemetry testing files
